feat(experimental): extract ExperimentalFeature type into standalone package
