+++ packages/suite/tsconfig.json
@@ -138,6 +138,7 @@
             "path": "../../suite-common/walletconnect"
         },
         { "path": "../../suite/analytics" },
+        { "path": "../../suite/experimental" },
         {
             "path": "../../suite/idb-migration-utils"
         },
