+++ suite/experimental/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@suite/experimental",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "depcheck": "yarn g:depcheck",
+        "type-check": "yarn g:tsc --build"
+    }
+}
