+++ suite/experimental/src/index.ts
@@ -1,0 +1,8 @@
+export type ExperimentalFeature =
+    | 'password-manager'
+    | 'tor-external'
+    | 'testnet-networks'
+    | 'nft-section'
+    | 'slip24'
+    | 'experimental-networks'
+    | 'suite-sync';
